promo order and stire

+++ lib/features/personal/post/domain/entities/post_entity.dart
@@ -116,7 +116,7 @@ class PostEntity {
   @HiveField(14)
   final DeliveryType deliveryType;
   @HiveField(15)
-  final List<int> listOfReviews;
+  final List<double> listOfReviews;
   @HiveField(16)
   final String categoryType;
